Support for scaling widgets.
This commit adds a new `scale_` property to the `Widget` class for the corresponded widget view to render scaled widgets.

The `WidgetView::PopulateWidgetList()` method is changed significantly to adapt widget's scale factor, and `Widget::Render()` calls `nvgScale()` to render scaled widgets.

The `Widget` class' `GetX()` and `GetY()` methods are changed to consider parent's scale factor as well. Further, `float` type is returned for all of `GetX()`, `GetY()`, `GetWidth()` and `GetHeight()` methods now because nanovg mainly accepts `float` type for rendering.

A minor change is made to `Widget::Redraw()` as well that it doesn't request the corresponded widget view to redraw immediately if the widget itself is hidden.

+++ moui/widgets/scroll_view.cc
@@ -197,10 +197,10 @@ void ScrollView::GetContentViewOriginLimits(int* minimum_x,
     *minimum_x = GetCurrentPage() * this->GetWidth() * -1;
     *maximum_x = *minimum_x;
   } else {
-    *minimum_x = std::min(0, this->GetWidth() - content_view_->GetWidth());
+    *minimum_x = std::min(0.0f, this->GetWidth() - content_view_->GetWidth());
     *maximum_x = 0;
   }
-  *minimum_y = std::min(0, this->GetHeight() - content_view_->GetHeight());
+  *minimum_y = std::min(0.0f, this->GetHeight() - content_view_->GetHeight());
   *maximum_y = 0;
 }
 
@@ -270,9 +270,8 @@ void ScrollView::HandleEvent(Event* event) {
   if (event->type() == Event::Type::kDown) {
     StopAnimation();
     event_history_.push_back({current_location, kTimestamp});
-    initial_scroll_content_view_origin_ = {
-        static_cast<float>(content_view_->GetX()),
-        static_cast<float>(content_view_->GetY())};
+    initial_scroll_content_view_origin_ = {content_view_->GetX(),
+                                           content_view_->GetY()};
     initial_scroll_page_ = GetCurrentPage();
   } else if (event->type() == Event::Type::kMove) {
     event_history_.push_back({current_location, kTimestamp});
@@ -569,8 +568,7 @@ void ScrollView::WidgetWillRender(NVGcontext* context) {
       kCurrentTimestamp - vertical_animation_state_.initial_timestamp;
 
   // Moves the content view based on the elapsed time.
-  Point origin = {static_cast<float>(content_view_->GetX()),
-                  static_cast<float>(content_view_->GetY())};
+  Point origin = {content_view_->GetX(), content_view_->GetY()};
   if (horizontal_animation_state_.is_animating)
     origin.x = GetCurrentAnimatingLocation(horizontal_animation_state_);
   if (vertical_animation_state_.is_animating)
